Deployments/grpc_server/requirements.txt
grpcio
grpcio-tools
pyyaml
torch
torchvision
